+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/indexes/TextIndexTest.java
@@ -139,8 +139,11 @@ import static com.apple.foundationdb.record.RecordCursor.NoNextReason.SOURCE_EXH
 import static com.apple.foundationdb.record.metadata.Key.Expressions.concat;
 import static com.apple.foundationdb.record.metadata.Key.Expressions.concatenateFields;
 import static com.apple.foundationdb.record.metadata.Key.Expressions.field;
+import static com.apple.foundationdb.record.metadata.Key.Expressions.recordType;
 import static com.apple.foundationdb.record.provider.foundationdb.indexes.TextIndexBunchedSerializerTest.entryOf;
 import static com.apple.foundationdb.record.provider.foundationdb.indexes.TextIndexTestUtils.COMPLEX_DOC;
+import static com.apple.foundationdb.record.provider.foundationdb.indexes.TextIndexTestUtils.MAP_DOC;
+import static com.apple.foundationdb.record.provider.foundationdb.indexes.TextIndexTestUtils.MULTI_DOC;
 import static com.apple.foundationdb.record.provider.foundationdb.indexes.TextIndexTestUtils.SIMPLE_DOC;
 import static com.apple.foundationdb.record.query.plan.match.PlanMatchers.bounds;
 import static com.apple.foundationdb.record.query.plan.match.PlanMatchers.coveringIndexScan;
@@ -163,9 +166,11 @@ import static org.hamcrest.Matchers.anyOf;
 import static org.hamcrest.Matchers.anything;
 import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.empty;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.greaterThanOrEqualTo;
+import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.lessThan;
@@ -211,10 +216,9 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     private static final Index MAP_ON_VALUE_PREFIX = new Index("Map$entry-value_prefix", new GroupingKeyExpression(field("entry", FanType.FanOut).nest(concatenateFields("key", "value")), 1), IndexTypes.TEXT,
             ImmutableMap.of(IndexOptions.TEXT_TOKENIZER_NAME_OPTION, PrefixTextTokenizer.NAME, IndexOptions.TEXT_TOKENIZER_VERSION_OPTION, "1"));
     private static final Index MAP_ON_VALUE_GROUPED_INDEX = new Index("Map$entry-value_by_group", new GroupingKeyExpression(concat(field("group"), field("entry", FanType.FanOut).nest(concatenateFields("key", "value"))), 1), IndexTypes.TEXT);
-    private static final String MAP_DOC = "MapDocument";
 
     @BeforeEach
-    public void resetRegistry() {
+    void resetRegistry() {
         TextTokenizerRegistryImpl.instance().reset();
     }
 
@@ -403,7 +407,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void saveSimpleDocuments() throws Exception {
+    void saveSimpleDocuments() throws Exception {
         final SimpleDocument simpleDocument = SimpleDocument.newBuilder()
                 .setDocId(1066L)
                 .setText("This is a simple document. There isn't much going on here, if I'm honest.")
@@ -521,7 +525,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     // be returned after the scan that were both greater than the map key due to a race condition.
     // This was able to reproduce the error when run alone.
     @Test
-    public void backwardsRangeScanRaceCondition() throws Exception {
+    void backwardsRangeScanRaceCondition() throws Exception {
         final Random r = new Random(0x5ca1ab1e);
         final List<String> lexicon = Arrays.asList(TextSamples.ROMEO_AND_JULIET_PROLOGUE.split(" "));
         final SimpleDocument bigDocument = getRandomRecords(r, 1, lexicon, 100, 0).get(0);
@@ -554,7 +558,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void saveSimpleDocumentsWithPrefix() throws Exception {
+    void saveSimpleDocumentsWithPrefix() throws Exception {
         final SimpleDocument shakespeareDocument = SimpleDocument.newBuilder()
                 .setDocId(1623L)
                 .setText(TextSamples.ROMEO_AND_JULIET_PROLOGUE)
@@ -593,7 +597,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void saveSimpleDocumentsWithFilter() throws Exception {
+    void saveSimpleDocumentsWithFilter() throws Exception {
         final SimpleDocument russianDocument = SimpleDocument.newBuilder()
                 .setDocId(1547L)
                 .setText(TextSamples.RUSSIAN)
@@ -625,7 +629,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void saveSimpleDocumentsWithSuffixes() throws Exception {
+    void saveSimpleDocumentsWithSuffixes() throws Exception {
         final SimpleDocument germanDocument = SimpleDocument.newBuilder()
                 .setDocId(1623L)
                 .setText(TextSamples.GERMAN)
@@ -655,7 +659,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void saveSimpleDocumentsWithNoPositions() throws Exception {
+    void saveSimpleDocumentsWithNoPositions() throws Exception {
         final SimpleDocument shakespeareDocument = SimpleDocument.newBuilder()
                 .setDocId(1623L)
                 .setText(TextSamples.ROMEO_AND_JULIET_PROLOGUE)
@@ -700,7 +704,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void saveSimpleDocumentsWithPositionsOptionChange() throws Exception {
+    void saveSimpleDocumentsWithPositionsOptionChange() throws Exception {
         final SimpleDocument shakespeareDocument = SimpleDocument.newBuilder()
                 .setDocId(1623L)
                 .setText(TextSamples.ROMEO_AND_JULIET_PROLOGUE)
@@ -749,7 +753,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void saveComplexDocuments() throws Exception {
+    void saveComplexDocuments() throws Exception {
         ComplexDocument complexDocument = ComplexDocument.newBuilder()
                 .setGroup(0)
                 .setDocId(1066L)
@@ -832,7 +836,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void saveComplexMultiDocuments() throws Exception {
+    void saveComplexMultiDocuments() throws Exception {
         final ComplexDocument shakespeareDocument = ComplexDocument.newBuilder()
                 .setGroup(0)
                 .setDocId(1623L)
@@ -871,7 +875,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void saveMapDocuments() throws Exception {
+    void saveMapDocuments() throws Exception {
         final MapDocument firstDocument = MapDocument.newBuilder()
                 .setDocId(1066L)
                 .addEntry(MapDocument.Entry.newBuilder()
@@ -928,7 +932,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void saveCombinedByGroup() throws Exception {
+    void saveCombinedByGroup() throws Exception {
         final SimpleDocument simpleDocument = SimpleDocument.newBuilder()
                 .setGroup(0)
                 .setDocId(1907L)
@@ -992,7 +996,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void saveSimpleWithAggressiveConflictRanges() throws Exception {
+    void saveSimpleWithAggressiveConflictRanges() throws Exception {
         // These two documents are from different languages and thus have no conflicts, so
         // without the aggressive conflict ranges, they wouldn't conflict if not
         // for the aggressive conflict ranges
@@ -1017,7 +1021,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void saveComplexWithAggressiveConflictRanges() throws Exception {
+    void saveComplexWithAggressiveConflictRanges() throws Exception {
         // These two documents are in different groups, so even with aggressive conflict
         // ranges, they should be able to be committed concurrently.
         final ComplexDocument zeroGroupDocument = ComplexDocument.newBuilder()
@@ -1039,7 +1043,167 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void tokenizerVersionChange() throws Exception {
+    void deleteWhereSimpleDocuments() {
+        final SimpleDocument doc1 = SimpleDocument.newBuilder()
+                .setDocId(1623L)
+                .setGroup(0)
+                .setText("foo bar")
+                .build();
+        final SimpleDocument doc2 = SimpleDocument.newBuilder()
+                .setDocId(1945L)
+                .setGroup(0)
+                .setText("foo bar")
+                .build();
+        final RecordMetaDataHook hook = metaDataBuilder -> {
+            TextIndexTestUtils.addRecordTypePrefix(metaDataBuilder);
+            // Add "text" to the simple document primary key so that when we issue a delete where based
+            // on the first field of the SimpleDocument$text index, it will align with the primary key
+            metaDataBuilder.getRecordType(SIMPLE_DOC)
+                    .setPrimaryKey(concat(recordType(), field("text"), field("doc_id")));
+        };
+        try (FDBRecordContext context = openContext()) {
+            openRecordStore(context, hook);
+            recordStore.saveRecord(doc1);
+            recordStore.saveRecord(doc2);
+            commit(context);
+        }
+        try (FDBRecordContext context = openContext()) {
+            openRecordStore(context, hook);
+            Query.InvalidExpressionException err = assertThrows(Query.InvalidExpressionException.class,
+                    () -> recordStore.deleteRecordsWhere(SIMPLE_DOC, Query.field("text").equalsValue("foo bar")));
+            assertThat(err.getMessage(), containsString("deleteRecordsWhere not supported by index SimpleDocument$text"));
+        }
+    }
+
+    @Test
+    void deleteWhereComplexByGroup() {
+        final ComplexDocument zeroGroupDocument = ComplexDocument.newBuilder()
+                .setGroup(0)
+                .setDocId(1623L)
+                .setText(TextSamples.ROMEO_AND_JULIET_PROLOGUE)
+                .build();
+        final ComplexDocument oneGroupDocument = ComplexDocument.newBuilder()
+                .setGroup(1)
+                .setDocId(1623L)
+                .setText(TextSamples.ROMEO_AND_JULIET_PROLOGUE)
+                .build();
+        final RecordMetaDataHook hook = metaDataBuilder -> {
+            TextIndexTestUtils.addRecordTypePrefix(metaDataBuilder);
+            metaDataBuilder.addIndex(COMPLEX_DOC, COMPLEX_TEXT_BY_GROUP);
+        };
+
+        try (FDBRecordContext context = openContext()) {
+            openRecordStore(context, hook);
+            recordStore.saveRecord(zeroGroupDocument);
+            recordStore.saveRecord(oneGroupDocument);
+            context.commit();
+        }
+
+        try (FDBRecordContext context = openContext()) {
+            openRecordStore(context, hook);
+
+            RecordQuery query = RecordQuery.newBuilder()
+                    .setRecordType(COMPLEX_DOC)
+                    .setFilter(Query.and(
+                            Query.field("group").equalsParameter("group_value"),
+                            Query.field("text").text().containsPhrase("nought could remove")
+                    ))
+                    .build();
+            RecordQueryPlan plan = recordStore.planQuery(query);
+
+            // Both groups should have a single element before delete where
+            assertThat(plan.execute(recordStore, EvaluationContext.forBinding("group_value", 0)).asList().join(), hasSize(1));
+            assertThat(plan.execute(recordStore, EvaluationContext.forBinding("group_value", 1)).asList().join(), hasSize(1));
+
+            recordStore.deleteRecordsWhere(COMPLEX_DOC, Query.field("group").equalsValue(0));
+
+            // Group zero should be gone but group 1 should still have an element remaining
+            assertThat(plan.execute(recordStore, EvaluationContext.forBinding("group_value", 0)).asList().join(), empty());
+            assertThat(plan.execute(recordStore, EvaluationContext.forBinding("group_value", 1)).asList().join(), hasSize(1));
+        }
+    }
+
+    @Test
+    void deleteWhereGroupedMap() {
+        MapDocument doc1 = MapDocument.newBuilder()
+                .setDocId(42)
+                .setGroup(1066L)
+                .addEntry(MapDocument.Entry.newBuilder()
+                        .setKey("foo")
+                        .setValue(TextSamples.ANGSTROM)
+                        .build())
+                .build();
+        MapDocument doc2 = MapDocument.newBuilder()
+                .setDocId(42)
+                .setGroup(1623L)
+                .addEntry(MapDocument.Entry.newBuilder()
+                        .setKey("foo")
+                        .setValue(TextSamples.ANGSTROM)
+                        .build())
+                .build();
+
+        final RecordMetaDataHook hook = metaDataBuilder -> {
+            TextIndexTestUtils.addRecordTypePrefix(metaDataBuilder);
+            final RecordTypeBuilder mapDoc = metaDataBuilder.getRecordType(MAP_DOC);
+            mapDoc.setPrimaryKey(concat(recordType(), field("group"), field("doc_id")));
+            metaDataBuilder.addIndex(mapDoc, MAP_ON_VALUE_GROUPED_INDEX);
+        };
+        try (FDBRecordContext context = openContext()) {
+            openRecordStore(context, hook);
+            recordStore.saveRecord(doc1);
+            recordStore.saveRecord(doc2);
+            commit(context);
+        }
+        try (FDBRecordContext context = openContext()) {
+            openRecordStore(context, hook);
+
+            RecordQuery query = RecordQuery.newBuilder()
+                    .setRecordType(MAP_DOC)
+                    .setFilter(Query.and(
+                            Query.field("group").equalsParameter("group_value"),
+                            Query.field("entry").oneOfThem().matches(Query.and(
+                                    Query.field("key").equalsValue("foo"),
+                                    Query.field("value").text().containsAll("unit named", 4)
+                            ))
+                    ))
+                    .setRemoveDuplicates(false)
+                    .build();
+            RecordQueryPlan plan = recordStore.planQuery(query);
+            assertThat(plan, textIndexScan(indexName(MAP_ON_VALUE_GROUPED_INDEX.getName())));
+
+            assertEquals(Collections.singletonList(doc1),
+                    plan.execute(recordStore, EvaluationContext.forBinding("group_value", doc1.getGroup()))
+                            .map(FDBQueriedRecord::getRecord)
+                            .map(rec -> MapDocument.newBuilder().mergeFrom(rec).build())
+                            .asList()
+                            .join());
+            assertEquals(Collections.singletonList(doc2),
+                    plan.execute(recordStore, EvaluationContext.forBinding("group_value", doc2.getGroup()))
+                            .map(FDBQueriedRecord::getRecord)
+                            .map(rec -> MapDocument.newBuilder().mergeFrom(rec).build())
+                            .asList()
+                            .join());
+
+            // Delete where for one of the two groups
+            recordStore.deleteRecordsWhere(MAP_DOC, Query.field("group").equalsValue(doc1.getGroup()));
+
+            assertEquals(Collections.emptyList(),
+                    plan.execute(recordStore, EvaluationContext.forBinding("group_value", doc1.getGroup()))
+                            .map(FDBQueriedRecord::getRecord)
+                            .map(rec -> MapDocument.newBuilder().mergeFrom(rec).build())
+                            .asList()
+                            .join());
+            assertEquals(Collections.singletonList(doc2),
+                    plan.execute(recordStore, EvaluationContext.forBinding("group_value", doc2.getGroup()))
+                            .map(FDBQueriedRecord::getRecord)
+                            .map(rec -> MapDocument.newBuilder().mergeFrom(rec).build())
+                            .asList()
+                            .join());
+        }
+    }
+
+    @Test
+    void tokenizerVersionChange() throws Exception {
         final SimpleDocument shakespeareDocument = SimpleDocument.newBuilder()
                 .setDocId(1623L)
                 .setText(TextSamples.ROMEO_AND_JULIET_PROLOGUE)
@@ -1112,7 +1276,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void tokenizerVersionChangeWithMultipleEntries() throws Exception {
+    void tokenizerVersionChangeWithMultipleEntries() throws Exception {
         final MapDocument map1 = MapDocument.newBuilder()
                 .setDocId(1066L)
                 .addEntry(MapDocument.Entry.newBuilder().setKey("fr").setValue(TextSamples.FRENCH))
@@ -1239,7 +1403,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
         }
     }
 
-    public void scanWithSkip(@Nonnull Index index, @Nonnull String token, int skip, int limit, boolean reverse) throws Exception {
+    void scanWithSkip(@Nonnull Index index, @Nonnull String token, int skip, int limit, boolean reverse) throws Exception {
         try (FDBRecordContext context = openContext()) {
             openRecordStore(context);
             final List<IndexEntry> fullResults = scanIndex(recordStore, index, TupleRange.allOf(Tuple.from(token)));
@@ -1266,7 +1430,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void scan() throws Exception {
+    void scan() throws Exception {
         final Random r = new Random(0x5ca1ab1e);
         List<SimpleDocument> records = getRandomRecords(r, 50);
         try (FDBRecordContext context = openContext()) {
@@ -1297,7 +1461,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void invalidScans() throws Exception {
+    void invalidScans() throws Exception {
         try (FDBRecordContext context = openContext()) {
             openRecordStore(context);
             final Index index = recordStore.getRecordMetaData().getIndex(TextIndexTestUtils.SIMPLE_DEFAULT_NAME);
@@ -1365,12 +1529,6 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
                 .get();
     }
 
-    @Nonnull
-    private List<Long> querySimpleDocumentsWithIndex(@Nonnull QueryComponent filter, int planHash, boolean isCoveringIndexExpected,
-                                                     @Nonnull Matcher<? super Comparisons.TextComparison> comparisonMatcher) throws InterruptedException, ExecutionException {
-        return querySimpleDocumentsWithIndex(filter, TextIndexTestUtils.SIMPLE_DEFAULT_NAME, planHash, isCoveringIndexExpected, comparisonMatcher);
-    }
-
     @Nullable
     private List<Long> querySimpleDocumentsWithIndex(@Nonnull QueryComponent filter, @Nonnull String indexName, @Nonnull QueryComponent textFilter, int planHash, boolean isCoveringIndexExpected) throws InterruptedException, ExecutionException {
         if (textFilter instanceof ComponentWithComparison && ((ComponentWithComparison)textFilter).getComparison() instanceof Comparisons.TextComparison) {
@@ -1402,7 +1560,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void querySimpleDocuments() throws Exception {
+    void querySimpleDocuments() throws Exception {
         final List<SimpleDocument> documents = TextIndexTestUtils.toSimpleDocuments(Arrays.asList(
                 TextSamples.ANGSTROM,
                 TextSamples.AETHELRED,
@@ -1518,7 +1676,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void queryDocumentsWithScanLimit() throws Exception {
+    void queryDocumentsWithScanLimit() throws Exception {
         // Load a big (ish) data set
         final int recordCount = 100;
         final int batchSize = 10;
@@ -1575,7 +1733,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void querySimpleDocumentsWithAdditionalFilters() throws Exception {
+    void querySimpleDocumentsWithAdditionalFilters() throws Exception {
         final List<SimpleDocument> documents = TextIndexTestUtils.toSimpleDocuments(Arrays.asList(
                 TextSamples.ROMEO_AND_JULIET_PROLOGUE,
                 TextSamples.ROMEO_AND_JULIET_PROLOGUE,
@@ -1679,7 +1837,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void querySimpleDocumentsWithDifferentTokenizers() throws Exception {
+    void querySimpleDocumentsWithDifferentTokenizers() throws Exception {
         final List<SimpleDocument> documents = TextIndexTestUtils.toSimpleDocuments(Arrays.asList(
                 TextSamples.ROMEO_AND_JULIET_PROLOGUE,
                 TextSamples.RUSSIAN,
@@ -1753,7 +1911,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void querySimpleDocumentsMaybeCovering() throws Exception {
+    void querySimpleDocumentsMaybeCovering() throws Exception {
         final List<SimpleDocument> documents = TextIndexTestUtils.toSimpleDocuments(Arrays.asList(
                 TextSamples.ANGSTROM,
                 TextSamples.AETHELRED,
@@ -1918,7 +2076,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void querySimpleDocumentsWithoutPositions() throws Exception {
+    void querySimpleDocumentsWithoutPositions() throws Exception {
         final List<SimpleDocument> documents = TextIndexTestUtils.toSimpleDocuments(Arrays.asList(
                 TextSamples.ANGSTROM,
                 TextSamples.AETHELRED,
@@ -2051,7 +2209,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void queryComplexDocuments() throws Exception {
+    void queryComplexDocuments() throws Exception {
         final List<String> textSamples = Arrays.asList(
                 TextSamples.ANGSTROM,
                 TextSamples.ROMEO_AND_JULIET_PROLOGUE,
@@ -2097,7 +2255,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void queryComplexDocumentsWithAdditionalFilters() throws Exception {
+    void queryComplexDocumentsWithAdditionalFilters() throws Exception {
         final List<String> textSamples = Arrays.asList(
                 TextSamples.ANGSTROM,
                 TextSamples.ROMEO_AND_JULIET_PROLOGUE,
@@ -2200,7 +2358,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void queryComplexDocumentsCovering() throws Exception {
+    void queryComplexDocumentsCovering() throws Exception {
         final List<String> textSamples = Arrays.asList(
                 TextSamples.FRENCH,
                 TextSamples.GERMAN,
@@ -2300,7 +2458,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void queryMultiTypeDocuments() throws Exception {
+    void queryMultiTypeDocuments() throws Exception {
         final List<String> bothTypes = Arrays.asList(SIMPLE_DOC, COMPLEX_DOC);
         final List<String> simpleTypes = Collections.singletonList(SIMPLE_DOC);
         final List<String> complexTypes = Collections.singletonList(COMPLEX_DOC);
@@ -2412,7 +2570,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void queryMapDocuments() throws Exception {
+    void queryMapDocuments() throws Exception {
         final List<String> textSamples = Arrays.asList(
                 TextSamples.ROMEO_AND_JULIET_PROLOGUE,
                 TextSamples.AETHELRED,
@@ -2507,7 +2665,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void queryMapsWithGroups() throws Exception {
+    void queryMapsWithGroups() throws Exception {
         final List<String> textSamples = Arrays.asList(
                 TextSamples.ROMEO_AND_JULIET_PROLOGUE,
                 TextSamples.AETHELRED,
@@ -2621,7 +2779,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
      */
     @MethodSource("indexArguments")
     @ParameterizedTest
-    public void queryScanEquivalence(@Nonnull Index index) throws Exception {
+    void queryScanEquivalence(@Nonnull Index index) throws Exception {
         final Random r = new Random(0xba5eba1L + index.getName().hashCode());
         final int recordCount = 100;
         final int recordBatch = 25;
@@ -2740,7 +2898,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void invalidQueries() {
+    void invalidQueries() {
         try (FDBRecordContext context = openContext()) {
             openRecordStore(context, metaDataBuilder -> metaDataBuilder.addIndex(SIMPLE_DOC, new Index("SimpleDocument$max_group", field("group").ungrouped(), IndexTypes.MAX_EVER_TUPLE)));
             List<RecordQuery> queries = Arrays.asList(
@@ -2787,7 +2945,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
     }
 
     @Test
-    public void invalidIndexes() {
+    void invalidIndexes() {
         final String testIndex = "test_index";
         List<KeyExpression> expressions = Arrays.asList(
                 // VersionKeyExpression.VERSION, // no version expression allowed
@@ -2804,7 +2962,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
         }
 
         // Verify doesn't work with repeated fields but in a case where a repeated field is possible.
-        invalidateIndex(KeyExpression.InvalidExpressionException.class, "MultiDocument",
+        invalidateIndex(KeyExpression.InvalidExpressionException.class, MULTI_DOC,
                 new Index(testIndex, field("text", FanType.FanOut), IndexTypes.TEXT));
 
         // Verify that unique indexes are invalid
@@ -2945,7 +3103,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
 
     @Tag(Tags.Performance)
     @Test
-    public void textIndexPerf1000SerialInsert() throws Exception {
+    void textIndexPerf1000SerialInsert() throws Exception {
         // Create 1000 records
         Random r = new Random();
         List<SimpleDocument> records = getRandomRecords(r, 1000);
@@ -2969,7 +3127,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
 
     @Tag(Tags.Performance)
     @Test
-    public void textIndexPerf100InsertOneBatch() throws Exception {
+    void textIndexPerf100InsertOneBatch() throws Exception {
         // Create 1000 records
         Random r = new Random();
         List<SimpleDocument> records = getRandomRecords(r, 100);
@@ -2991,7 +3149,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
 
     @Tag(Tags.Performance)
     @Test
-    public void textIndexPerf1000SerialInsertNoBatching() throws Exception {
+    void textIndexPerf1000SerialInsertNoBatching() throws Exception {
         // Create 1000 records
         Random r = new Random();
         List<SimpleDocument> records = getRandomRecords(r, 1000);
@@ -3013,7 +3171,7 @@ public class TextIndexTest extends FDBRecordStoreTestBase {
 
     @Tag(Tags.Performance)
     @Test
-    public void textIndexPerf1000ParallelInsert() throws Exception {
+    void textIndexPerf1000ParallelInsert() throws Exception {
         // Create 1000 records
         Random r = new Random();
         List<SimpleDocument> records = getRandomRecords(r, 1000);